# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Poll": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"closed_at": {
						"type": "string",
						"format": "date-time"
					},
					"question": {
						"type": "string"
					},
					"options": {
						"items": {
							"$ref": "#/components/schemas/Option"
						},
						"type": "array"
					},
					"voters": {
						"additionalProperties": false,
						"type": "object"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"active",
					"created_at",
					"closed_at",
					"question",
					"options",
					"voters",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Option": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"votes": {
						"type": "number",
						"format": "double"
					},
					"text": {
						"type": "string"
					},
					"poll": {
						"$ref": "#/components/schemas/Poll"
					}
				},
				"required": [
					"id",
					"votes",
					"text",
					"poll"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"polls": {
						"items": {
							"$ref": "#/components/schemas/Poll"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password",
					"polls"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"polls": {
						"items": {
							"$ref": "#/components/schemas/Poll"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"polls"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPollPayload": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"closed_at": {
						"type": "string",
						"format": "date-time"
					},
					"question": {
						"type": "string"
					},
					"options": {
						"items": {
							"$ref": "#/components/schemas/Option"
						},
						"type": "array"
					},
					"voters": {
						"additionalProperties": false,
						"type": "object",
						"nullable": true
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"active",
					"created_at",
					"closed_at",
					"question",
					"options",
					"voters",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "db_fiddling",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/users/{email}": {
			"get": {
				"operationId": "GetUserByEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/polls": {
			"get": {
				"operationId": "GetPolls",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Poll"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Poll"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePoll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Poll"
								}
							}
						}
					}
				},
				"tags": [
					"Poll"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPollPayload"
							}
						}
					}
				}
			}
		},
		"/polls/{id}": {
			"get": {
				"operationId": "GetPoll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Poll"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Poll"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/option/{id}": {
			"get": {
				"operationId": "CastVote",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Option"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}